fix(a11y): SSR the landing without cookies() to keep the static export building
The cookie-hint variant made `/` a server component reading `cookies()`, which
is a dynamic API incompatible with the mobile/Capacitor `output: export` build
(Mobile APK Build failed). Drop it: `/` is a client component again that renders
the landing on its first (server) pass instead of `null`, so the SSR HTML still
carries the landing's <main> + <h1> (A11Y-001/002 + LH-A11Y) and the static
export keeps working. Trade-off: authenticated users briefly see the landing
during the silent refresh before the dashboard swaps in (documented).
