+++ pwa/src/app/page.tsx
@@ -1,20 +1,57 @@
-import { cookies } from "next/headers";
-import { HomeContent } from "@/components/home-content";
+"use client";
+
+import { Suspense, useEffect, useRef } from "react";
+import { useAuthStore } from "@/store/auth-store";
+import { HydrationBoundary } from "@/components/hydration-boundary";
+import { TripPlanner } from "@/components/trip-planner";
+import { TripPlannerErrorBoundary } from "@/components/trip-planner-error-boundary";
+import { LandingPage } from "@/components/landing-page";
 
 /**
  * Home page (dual-state: anonymous landing / authenticated dashboard).
  *
- * Server component so the anonymous landing is server-rendered with its `<main>`
- * and `<h1>` in the initial HTML (audit 35.2 A11Y-001/002 + LH-A11Y) — crawlable
- * and accessible, instead of the previous client-only render that emitted an
- * empty document. The presence of the httpOnly `refresh_token` cookie is passed
- * as a hint so authenticated users render the dashboard directly (no landing
- * flash) without a hydration mismatch; the client silent refresh then confirms.
+ * `/` is public, so it runs its own silent auth check on mount. The landing is
+ * rendered on the FIRST (server) pass — instead of the previous `null` — so the
+ * SSR HTML carries the landing's `<main>` + `<h1>` (audit 35.2 A11Y-001/002 +
+ * LH-A11Y, and a better LCP). Authenticated users briefly see the landing during
+ * the silent refresh, then swap to the dashboard. (A `cookies()`-based hint would
+ * avoid that flash but makes the route dynamic, which is incompatible with the
+ * static mobile/Capacitor `output: export` build.)
  */
-export default async function Page() {
-  const cookieStore = await cookies();
-  // Cookie name mirrors api `AuthCookies::REFRESH_TOKEN`.
-  const authHint = cookieStore.get("refresh_token")?.value ? "authed" : "anon";
+function HomeContent() {
+  const isAuthenticated = useAuthStore((s) => s.isAuthenticated);
+  const checkStarted = useRef(false);
 
-  return <HomeContent authHint={authHint} />;
+  useEffect(() => {
+    if (checkStarted.current) return;
+    checkStarted.current = true;
+    if (isAuthenticated) return;
+    void useAuthStore.getState().silentRefresh();
+  }, [isAuthenticated]);
+
+  if (!isAuthenticated) {
+    return (
+      <Suspense fallback={null}>
+        <LandingPage />
+      </Suspense>
+    );
+  }
+
+  return (
+    <HydrationBoundary>
+      <TripPlannerErrorBoundary>
+        <Suspense fallback={null}>
+          <TripPlanner />
+        </Suspense>
+      </TripPlannerErrorBoundary>
+    </HydrationBoundary>
+  );
+}
+
+export default function Page() {
+  return (
+    <Suspense fallback={null}>
+      <HomeContent />
+    </Suspense>
+  );
 }
